Use Call in service, change timestamp to int for seconds

+++ data/src/main/java/org/oppia/android/data/backends/gae/model/GaeFeedbackReport.kt
@@ -11,7 +11,7 @@ import com.squareup.moshi.JsonClass
 data class GaeFeedbackReport(
 
   /** The date and time in sec that the report was created. */
-  @Json(name = "report_submission_timestamp_sec") val reportSubmissionTimestampSec: Long,
+  @Json(name = "report_submission_timestamp_sec") val reportSubmissionTimestampSec: Int,
   /** The information that user's provide in the feedback report. */
   @Json(name = "user_supplied_feedback") val userSuppliedFeedback: GaeUserSuppliedFeedback,
   /** Information collected about the user device's system. */
